requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0
svglib>=1.5.1
reportlab>=4.0.4
